pyproject.toml
[project]
name = "torc"
version = "0.1.0"
description = ""
authors = [
    {name = "mateusb12", email = "matbessa12@gmail.com"}
]
requires-python = ">=3.12"

dependencies = [
    "fastapi (>=0.128.6,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.1,<8.0.0)",
    "repomix (>=0.4.1,<0.5.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
